requirements.txt
solana
requests
tqdm
google-cloud-storage
